% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.r
\name{add_max_feature}
\alias{add_max_feature}
\title{Assign samples to a new group variable based on the highest feature}
\usage{
add_max_feature(data, group = ".group", reorder.samples = TRUE)
}
\arguments{
\item{data}{\code{ExpressionSet} object.}

\item{group}{name of new grouping variable.}

\item{reorder.samples}{If \code{TRUE}, samples are reordered by the
prevalance of their assigned group and rank of the assigned feature's value
within the group.}
}
\description{
Assign samples to a new group variable based on the highest feature
}
\examples{
profiles <- add_max_feature(profiles, group = "enterotype")
profiles$enterotype

}

