% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.r
\name{assign_max_group}
\alias{assign_max_group}
\title{Create a new group variable based on the highest feature}
\usage{
assign_max_group(data, sample, feature, value, group = ".group")
}
\arguments{
\item{data}{\code{data.frame} or \code{tbl_df} containing relevant data}

\item{sample}{sample identifier column}

\item{feature}{feature identifier column}

\item{value}{column that stores values}

\item{group}{name of new grouping variable}
}
\description{
Each \code{sample} is assigned to a new grouping variable (default:
\code{.group}) named for the \code{feature} with the maximum \code{value}.
Furthermore, the \code{sample} column is returned as a factor ordered by the
prevalence of its assigned group and rank of the assigned \code{feature}'s
\code{value} among other samples in that group.
}
\examples{
assign_max_group(profiles, 'sample', 'taxa', 'abundance')

}

