% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.r
\name{cluster_dimensions}
\alias{cluster_dimensions}
\alias{cluster_features}
\alias{cluster_features.ExpressionSet}
\alias{cluster_samples}
\alias{cluster_samples.ExpressionSet}
\title{Reorder samples and features by hierarchical clustering}
\usage{
cluster_samples(data, dist.method = "euclidean", hclust.method = "average")

\method{cluster_samples}{ExpressionSet}(data, dist.method = "euclidean",
  hclust.method = "average")

cluster_features(data, dist.method = "euclidean", hclust.method = "average")

\method{cluster_features}{ExpressionSet}(data, dist.method = "euclidean",
  hclust.method = "average")
}
\arguments{
\item{data}{\code{ExpressionSet} object.}

\item{dist.method}{distance metric to be calculated prior to clustering.}

\item{hclust.method}{the agglomeration method to be used.}
}
\value{
An \code{ExpressionSet} object reordered on one dimension.
}
\description{
Reorder samples or features of an \code{ExpressionSet} object based on
agglomerative hierarchical clustering of their profiles.
}
\details{
Distance among samples or features is first calculated using the method
specified by \code{dist.method}, which can be any of the measures accepted by
\code{\link[stats]{dist}}. Similarly, \code{hclust.method} can be any of the
agglomeration methods accepted by \code{\link[stats]{hclust}}.
}
\examples{
cluster_samples(profiles)
cluster_features(profiles)

}
\seealso{
\code{\link[stats]{dist}}, \code{\link[stats]{hclust}}
}

