% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframes.r
\name{from_dataframe}
\alias{from_dataframe}
\title{Convert a tidy data.frame to an ExpressionSet object}
\usage{
from_dataframe(data, sample, feature, value, pvars = NULL, fvars = NULL)
}
\arguments{
\item{data}{tidy data.frame to convert.}

\item{sample}{name of column that stores sample identifiers}

\item{feature}{name of column that stores feature identifiers}

\item{value}{name of column that stores values}

\item{pvars}{phenotype variables, names of columns that contain metadata
associated with samples}

\item{fvars}{feature variables, names of columns that contains metadata
associated with features}
}
\description{
Convert a tidy data.frame to an ExpressionSet object
}
