% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.r
\name{order_by_prevalence}
\alias{order_by_prevalence}
\title{Order a variable by its prevalence in the data}
\usage{
order_by_prevalence(data, variable, direction = 1)
}
\arguments{
\item{data}{\code{data.frame} or \code{tbl_df} containing relevant data}

\item{variable}{with \code{data} to be ordered}

\item{direction}{determines ordering direction. If 1, the default, the
variable is ordered by descending prevalence.  If -1, the order is reversed.}
}
\description{
The order of the data is not actually changed. Rather, the specified variable
is converted to a factor ordered by the number of times it is observed in the
data.
}

