% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.r
\name{order_by_rank}
\alias{order_by_rank}
\title{Convert variables to ordered factors}
\usage{
order_by_rank(data, variable, value, by = NULL, fun.aggregate = NULL,
  direction = 1)
}
\arguments{
\item{data}{\code{data.frame} or \code{tbl_df} containing relevant data}

\item{variable}{with \code{data} to be ordered}

\item{value}{column that stores values}

\item{by}{a column name in \code{data}. If supplied, \code{value} rankings
are assigned within each level of \code{by}.}

\item{fun.aggregate}{function that takes a
vector of numbers and returns a single number; the default is \code{mean()}.}

\item{direction}{determines ordering direction. If 1, the default, the
variable is ordered by descending prevalence.  If -1, the order is reversed.}
}
\description{
Convert \code{variable} to a factor ordered by the rank of \code{value}. If
\code{variable} does not uniquely identify a value, an aggregating function,
\code{fun.aggregate}, is applied.
}
\details{
A grouping variable passed to \code{by} will cause rankings to be assigned
within each level of the grouping variable.
}
\examples{
order_by_rank(profiles, "taxa", "abundance")
}

