% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.r
\name{palettes}
\alias{brewer_pal_plus}
\alias{manual_pal_plus}
\alias{palettes}
\title{Augmented color palette functions}
\usage{
brewer_pal_plus(palette = 1, direction = 1)

manual_pal_plus(values)
}
\arguments{
\item{palette}{If a string, will use that named palette.  If a number, will
index into the list of palettes of appropriate \code{type}}

\item{direction}{Sets the order of colors in the scale. If 1, the default,
colors are as output by \code{\link[RColorBrewer]{brewer.pal}}. If -1, the
order of colors is reversed.}

\item{values}{vector of values to be used as a palette.}
}
\description{
Palette functions return a function that generates colors. These
\code{_plus()} functions are designed to work like their counterparts in the
\code{scales} package, with the exception that additional standard colors are
appended to the original palettes. The additional colors are not pretty but
some effort has been made to avoid overly similar colors.
}
\details{
The \code{manual_pal_plus()} is useful mapping specific colors to certain
features.
}

