% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-profile.r
\name{plot_profile}
\alias{plot_profile}
\title{Stacked bar plot of profile samples by features}
\usage{
plot_profile(data, sample, feature, value, legend = FALSE)
}
\arguments{
\item{data}{\code{data.frame} or \code{tbl_df} containing relevant data}

\item{sample}{sample identifier column}

\item{feature}{feature identifier column}

\item{value}{column that stores values}

\item{legend}{include a legend?}
}
\description{
Stacked bar plot of profile samples by features
}
\examples{
plot_profile(profiles, 'sample', 'taxa', 'abundance')

}

