% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-barplot.r
\name{profile_barplot}
\alias{profile_barplot}
\title{Stacked bar plot of sample feature profiles}
\usage{
profile_barplot(data, legend = FALSE)
}
\arguments{
\item{data}{\code{ExpressionSet} object.}

\item{legend}{include a legend?}
}
\description{
Stacked bar plot of sample feature profiles
}
\examples{
profile_barplot(profiles)

}

