% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-dotplot.r
\name{profile_dotplot}
\alias{profile_dotplot}
\title{Dotplot of sample feature profiles}
\usage{
profile_dotplot(data, color = NULL, palette, legend = TRUE, top.n = NULL,
  other.color = "grey50", max.size = 6, title.area, title.color)
}
\arguments{
\item{data}{\code{ExpressionSet} object.}

\item{color}{color code dots by sample, feature or any of the associated metadata variables.}

\item{palette}{a function that generates \code{n} colors.}

\item{legend}{include a legend?}

\item{top.n}{Limit the number of levels in the new grouping variable to the
\code{top.n} most prevalent. Samples not among the most prevalent groups
are assigned to \code{"Othere"}.}

\item{other.color}{Color applied to features not among the \code{top.n}.}

\item{max.size}{Size of largest dots.}

\item{title.area}{Title of point area legend.}

\item{title.color}{Title of point color legend.}
}
\description{
Dot size is determined by mapping point area to feature values. Dots can be
color coded by sample (\code{color = "sample"}), by feature (\code{color =
"feature"}) or any associated metadatda variable.
}
\examples{
profile_dotplot(profiles)
}
