% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{profiles}
\alias{profiles}
\title{Human Microbiome Project Metagenome Taxonomic Profiles}
\format{A dataset with variables
\describe{
\item{taxa}{species-level clade assigned by MetaPhlan2}
\item{sample}{Sequenced body site}
\item{abundance}{relative abundance of detected microbe}
}}
\source{
\url{http://huttenhower.sph.harvard.edu/metaphlan}
}
\usage{
profiles
}
\description{
Profiles of 51 vaginal metagenomic samples from the Human Microbiome Project and analyzed using MetaPhlan2.
}
\keyword{datasets}
