% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.r
\name{read_metaphalan}
\alias{read_metaphalan}
\title{Read metaphalan results}
\usage{
read_metaphalan(file, taxa.level = "species")
}
\arguments{
\item{file}{the name of the file which the data are to be read from.}

\item{taxa.level}{the taxonomic level for which the relative abundance output will be loaded.}
}
\value{
ExpressionSet
}
\description{
Read metaphalan results
}
\examples{
\dontrun{
prof.file <- system.file("inst/extdata/profiles.txt", package = "metafiler")
read_metaphalan(prof.file)
}
}

