% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.r
\name{reorder_dimensions}
\alias{reorder_dimensions}
\alias{reorder_samples}
\alias{reorder_samples.ExpressionSet}
\alias{reorder_features}
\alias{reorder_features.ExpressionSet}
\title{Reorder samples and features}
\usage{
reorder_samples(data, fun.aggregate = NULL, decreasing = TRUE)

\method{reorder_samples}{ExpressionSet}(data, fun.aggregate = NULL,
  decreasing = TRUE)

reorder_features(data, fun.aggregate = NULL, decreasing = TRUE)

\method{reorder_features}{ExpressionSet}(data, fun.aggregate = NULL,
  decreasing = TRUE)
}
\arguments{
\item{data}{\code{ExpressionSet} object.}

\item{fun.aggregate}{function that takes a vector of numbers and returns a
single number (defaults to \code{mean}).}

\item{decreasing}{logical. Should the sort be increasing or decreasing?}
}
\value{
An \code{ExpressionSet} object reordered on one dimension.
}
\description{
Reorder samples or features of an \code{ExpressionSet} object based on the
rank of their aggregated values. The manner in which values are aggregated is
determined by \code{fun.aggregate}, which defaults to \code{mean}.
}
\examples{
reorder_samples(profiles)
reorder_features(profiles)
}
