% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.r
\name{theme_metafiler}
\alias{theme_metafiler}
\title{Default profile barplot theme}
\usage{
theme_metafiler(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
Default profile barplot theme
}
