% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.r
\name{to_dataframe}
\alias{to_dataframe}
\title{Convert an ExpressionSet object to a tidy data.frame}
\usage{
to_dataframe(data, add.pvars = TRUE, add.fvars = TRUE)
}
\arguments{
\item{data}{an \code{ExpressionSet} object}

\item{add.pvars}{should phenotype variables stored in \code{pData(data)} be included in the \code{data.frame}?}

\item{add.fvars}{should feature variables stored in \code{fData(data)} be included in the \code{data.frame}?}
}
\description{
Convert an ExpressionSet object to a tidy data.frame
}

